including result enum, and Appdelegate minor fixes
