+++ BTGTesteChallenge/BTGTesteChallenge/ViewController.swift
@@ -13,6 +13,7 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
+        print("teste")
     }
